external/atria_types/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "atria_types"
version = "0.0.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE.txt" }
authors = [{ name = "Saifullah", email = "saifullah.saifullah@dfki.de" }]
dependencies = [
    "lazy_loader>=0.4",
    "numpy>=1.26.4",
    "omegaconf>=2.3.0",
    "Pillow>=11.0.0",
    "rich>=14.0.0",
    "beautifulsoup4>=4.12.2",
    "imagesize>=1.4.1",
    "pyarrow>=20.0.0",
    "pydantic>=2.11.7",
    "codename>=1.1",
    "lxml>=6.0.0",
    "pypdf2>=3.0.1",
    "pymupdf>=1.26.3",
    "pdf2image>=1.17.0",
    "requests>=2.32.5",
    "atria-logger",
    "factory-boy>=3.3.3",
    "torch==2.1.2",
    "torchvision==0.16.2",
]

[project.optional-dependencies]
test = ["pytest", "coverage"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
target-version = "py311"
extend-exclude = ["alembic"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP"]
ignore = [
    "E501",  # line too long, handled by formatter
    "D100", "D104",  # docstring warnings
]

[tool.ruff.format]
docstring-code-format = true

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-rXs --strict-config --strict-markers --tb=short --import-mode=importlib"

[project.urls]
Homepage = "https://github.com/saifullah3396/atria_types/"
Source = "https://github.com/saifullah3396/atria_types/"
BugReports = "https://github.com/saifullah3396/atria_types/"

[tool.mypy]
ignore_missing_imports = true
strict = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_unused_ignores = true

[tool.uv.sources]
atria-logger = { path = "../atria_logger", editable = true }

[dependency-groups]
dev = [
    "coverage>=7.12.0",
    "pytest>=9.0.1",
    "ruff>=0.14.6",
    "mypy>=1.18.2",
]
